mm-db-manager/cloud/pyproject.toml
[project]
name = "cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "functions-framework>=3.10.0",
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "load-dotenv>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
  "integration: tests that hit real external services (Sheets/HTTP)",
]
addopts = "-q"

[tool.pytest_env]
env_files = [".env"]
